My C - Functions, nested loops Task

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -14,7 +14,7 @@ int main(void)
 	unsigned long x = 0;
 	unsigned long y = 1;
 	unsigned long Sum;
-	unsigned long Total_Sum;
+	unsigned Total_Sum;
 
 	while (x <= 4000000)
 	{
@@ -28,7 +28,7 @@ int main(void)
 		y = Sum;
 	}
 
-	printf("%lu\n", Total_Sum);
+	printf("%u\n", Total_Sum);
 
 	return (0);
 }
